app/requirements.txt
sv-ttk==2.6.0
PyBluez==0.30
pyserial==3.5